Rename views with `View` postfix
Lint rules require all components be multi-word, so the View prefix
ensures all views meet the requirement, plus makes them more
greppable and consistent, even if they were already multi-word
