+++ packages/client/src/views/RecentActivityView.vue
